requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
django-oauth-toolkit==2.2.0
mozilla-django-oidc==3.0.0
django-filter==23.3
whitenoise==6.5.0
gunicorn==21.2.0
python-dotenv==1.0.0